Update speakers list style
